+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+## 1.21.3
+* Fixed `chatNotification` flag support for Android
+
 ## 1.21.2
 * Added `chatNotification` flag support for Android
+* Updated native iOS + Android SDKs
 
 ## 1.21.1
 * Fixed `trackEvent` for Android and support `properties` argument